+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Log unmapped names only if `DEBUG` logging is enabled.
+- Numerous documentation fixups and changes.
+
 ## [0.15.3] - 2025-10-19
 
 ### Fixed