+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.15.4] - 2025-10-25
+
 ### Changed
 - Log unmapped names only if `DEBUG` logging is enabled.
 - Numerous documentation fixups and changes.
@@ -459,7 +461,8 @@ cookiecutter template.
 - Switch to relative imports.
 - Fix some intersphinx issues.
 
-[Unreleased]: https://github.com/rsundqvist/id-translation/compare/v0.15.3...HEAD
+[Unreleased]: https://github.com/rsundqvist/id-translation/compare/v0.15.4...HEAD
+[0.15.4]: https://github.com/rsundqvist/id-translation/compare/v0.15.3...v0.15.4
 [0.15.3]: https://github.com/rsundqvist/id-translation/compare/v0.15.2...v0.15.3
 [0.15.2]: https://github.com/rsundqvist/id-translation/compare/v0.15.1...v0.15.2
 [0.15.1]: https://github.com/rsundqvist/id-translation/compare/v0.15.0...v0.15.1